rename webcomponents-loader to webcomponents-es6-loader
