+++ README.md
@@ -30,8 +30,8 @@ different browsers and spec readiness:
 If you are only targeting a specific browser, you can just use the bundle that's
 needed by it; alternatively, if your server is capable of serving different assets based on user agent, you can send the polyfill bundle that's necessary for the browser making that request.
 
-## `webcomponents-loader.js`
-Alternatively, this repo also comes with `webcomponents-loader.js`, a client-side
+## `webcomponents-es6-loader.js`
+Alternatively, this repo also comes with `webcomponents-es6-loader.js`, a client-side
 loader that dynamically loads the minimum polyfill bundle, using feature detection.
 Note that because the bundle will be loaded asynchronously, you should wait for the `WebComponentsReady` before you can safely assume that all the polyfills have
 loaded and are ready to be used (i.e. if you want to dynamically load other custom
@@ -39,7 +39,7 @@ elements, etc.). Here's an example:
 
 ```html
 <!-- Load polyfills; note that "loader" will load these async -->
-<script src="bower_components/webcomponentsjs/webcomponents-loader.js"></script>
+<script src="bower_components/webcomponentsjs/webcomponents-es6-loader.js"></script>
 
 <!-- Load a custom element definition via HTMLImports -->
 <link rel="import" href="my-element.html">
@@ -63,12 +63,12 @@ elements, etc.). Here's an example:
 ## `webcomponents-es5-loader.js`
 According to the spec, Custom Elements must be ES6 classes (https://html.spec.whatwg.org/multipage/scripting.html#custom-element-conformance). Since most projects need to support a wide range of browsers that don't necessary support ES6, it may make sense to compile your project to ES5. However, ES5-style custom element classes will **not** work with native Custom Elements because ES5-style classes cannot properly extend ES6 classes, like `HTMLElement`.
 
-To work around this, `webcomponents-es5-loader.js` first loads an extra ES5 compatibility [shim](https://github.com/webcomponents/custom-elements/blob/master/src/native-shim.js) before
-loading the minimum polyfill bundle, like `webcomponents-loader.js` does. 
+To work around this, `webcomponents-es5-loader.js` first executes an extra ES5 compatibility [shim](https://github.com/webcomponents/custom-elements/blob/master/src/native-shim.js) before
+loading the minimum polyfill bundle, like `webcomponents-es6-loader.js` does. 
 
-The reason for having two different loaders (`webcomponents-loader.js` for ES6 deployed code and `webcomponents-es5-loader.js` for ES5) is that the polyfill can't know if you're deploying ES5 or ES6 code -- it runs before your code, and only _you_ as an author of the code know what you're going to deploy. This means that _you_ have to make the decision about which loader to use. 
+The reason for having two different loaders (`webcomponents-es6-loader.js` for ES6 deployed code and `webcomponents-es5-loader.js` for ES5) is that the polyfill can't know if you're deploying ES5 or ES6 code -- it runs before your code, and only _you_ as an author of the code know what you're going to deploy. This means that _you_ have to make the decision about which loader to use. 
 **TL; DR: ** Use `webcomponents-es5-loader.js` only if you want to deploy ES5 code to _all_ browsers (both with and without native custom elements support) -- if this isn't a requirement, then use
-`webcomponents-loader.js` instead.
+`webcomponents-es6-loader.js` instead.
 
 ## Browser Support
