+++ tests/ce-import-upgrade.html
@@ -13,7 +13,7 @@
     <title></title>
     <meta charset="UTF-8">
     <script src="../../web-component-tester/browser.js"></script>
-    <script src="../webcomponents-loader.js"></script>
+    <script src="../webcomponents-es6-loader.js"></script>
     <link rel="import" href="imports/a1-import.html">
   </head>
   <body>
