+++ tests/ce-upgrade-order.html
@@ -12,7 +12,7 @@
 <head>
   <meta charset="UTF-8">
   <title>Custom Element Upgrade Order</title>
-  <script src="../webcomponents-loader.js"></script>
+  <script src="../webcomponents-es6-loader.js"></script>
   <script src="../../web-component-tester/browser.js"></script>
   <link rel="import" href="imports/import-upgrade-order.html">
 </head>
