+++ tests/ce-upgradedocumenttree.html
@@ -20,7 +20,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
       }
     </script>
     <script src="../../web-component-tester/browser.js"></script>
-    <script src="../webcomponents-loader.js"></script>
+    <script src="../webcomponents-es6-loader.js"></script>
     <link rel="import" href="imports/element-import-a.html">
   </head>
   <body>
