+++ tests/integration.html
@@ -11,7 +11,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 <html>
   <head>
     <title>Integration Test</title>
-    <script src="../webcomponents-loader.js"></script>
+    <script src="../webcomponents-es6-loader.js"></script>
     <script src="../../web-component-tester/browser.js"></script>
     <link rel="import" href="imports/simple-element.html">
   </head>
