+++ tests/load.html
@@ -21,7 +21,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
         window.webComponentsReadyCount++;
       });
     </script>
-    <script src="../webcomponents-loader.js"></script>
+    <script src="../webcomponents-es6-loader.js"></script>
     <script src="../../web-component-tester/browser.js"></script>
     <link rel="import" href="imports/import-file.html">
   </head>
