+++ tests/smoke.html
@@ -13,7 +13,7 @@
     <title></title>
     <meta charset="UTF-8">
     <script src="../../web-component-tester/browser.js"></script>
-    <script src="../webcomponents-loader.js"></script>
+    <script src="../webcomponents-es6-loader.js"></script>
   </head>
   <body>
     <x-foo>plain</x-foo>
